Disabled custom rtl8192cu in favor of rtlwifi
While we don't know why rtl8192cu got included in the first place, it does not
seem to work at all with Edimax dongle which is one of the more popular options
for use with Pi. On the other hand, rtlwifi driver collection which comes from
the mainstream kernel seems to work well (at least on Pi B+).